Update to 2.2.38, set minimum purchase amount to $0.50

+++ assets/js/components/billing/PurchaseCreditModal.jsx
@@ -602,7 +602,7 @@ class PurchaseCreditModal extends Component {
       ];
 
     const stripe_min_purchase = window.stripe_minimum_purchase
-      ? parseInt(window.stripe_minimum_purchase)
+      ? parseFloat(window.stripe_minimum_purchase)
       : 10;
     const allButtons = [
       <Button
